requirements.txt
# ── API & HTTP ─────────────────────────────────────────────────────────────
requests==2.32.3
urllib3==2.2.3

# ── Data Processing ────────────────────────────────────────────────────────
pandas==2.1.4
pyarrow==17.0.0          # Parquet read/write for S3 processed layer

# ── AWS ────────────────────────────────────────────────────────────────────
boto3==1.35.36
s3fs==2024.9.0           # Pandas S3 read/write support

# ── Snowflake ──────────────────────────────────────────────────────────────
snowflake-connector-python==3.12.3
snowflake-sqlalchemy==1.6.1

# ── Airflow ────────────────────────────────────────────────────────────────
apache-airflow==2.10.2
apache-airflow-providers-snowflake==5.7.1
apache-airflow-providers-amazon==8.28.0

# ── Environment & Config ───────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Testing ────────────────────────────────────────────────────────────────
pytest==8.3.3
pytest-mock==3.14.0

# ── Linting (optional but professional) ───────────────────────────────────
ruff==0.6.9